+++ lib/ooxml_parser/common_parser/common_data/ooxml_document_object.rb
@@ -51,8 +51,6 @@ module OoxmlParser
     class << self
       # @return [String] path to root subfolder
       attr_accessor :root_subfolder
-      # @return [PresentationTheme] list of themes
-      attr_accessor :theme
       # @return [Array<String>] stack of xmls
       attr_accessor :xmls_stack
       # @return [String] path to root folder