+++ lib/ooxml_parser/xlsx_parser/workbook.rb
@@ -123,7 +123,7 @@ module OoxmlParser
       OOXMLDocumentObject.root_subfolder = 'xl/'
       OOXMLDocumentObject.add_to_xmls_stack('xl/workbook.xml')
       @doc = Nokogiri::XML.parse(File.open(OOXMLDocumentObject.current_xml))
-      @theme = PresentationTheme.parse("xl/#{link_to_theme_xml}") if link_to_theme_xml
+      @theme = PresentationTheme.new.parse("xl/#{link_to_theme_xml}") if link_to_theme_xml
       @style_sheet = StyleSheet.new(parent: self).parse
       @doc.xpath('xmlns:workbook/xmlns:sheets/xmlns:sheet').each do |sheet|
         @sheets << Sheet.new(parent: self).parse(sheet)